README.md: change sed command to bypass update passwd dialog
If the root user has a passwd update pending, changing only the 'auth sufficient' line will bypass the initial verification, but the user will be trapped by the update passwd dialog, which asks for the old one. To avoid this, overwriting the 'password required' line is necessary, so the passwd update dialog is also bypassed.